requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
beanie==1.29.0
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.11
h11==0.14.0
idna==3.10
lazy-model==0.2.0
motor==3.7.0
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pymongo==4.11.3
python-dotenv==1.1.0
python-json-logger==3.3.0
sniffio==1.3.1
starlette==0.46.1
toml==0.10.2
typing_extensions==4.12.2
uvicorn==0.34.0
